+++ p2panda/tests/api.rs
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+use std::time::Duration;
+
 use futures_util::StreamExt;
-use p2panda::streams::{EphemeralMessage, ProcessedOperation, StreamEvent};
+use p2panda::streams::{EphemeralMessage, Offset, ProcessedOperation, StreamEvent};
 use p2panda::test_utils::setup_logging;
 use p2panda_core::{PrivateKey, Topic};
 use tokio::task::JoinHandle;
@@ -93,4 +95,48 @@ async fn eventually_consistent_stream() {
     let received = received.expect("icebear should have received operation");
     assert_eq!(received.message(), &"Hello, Icebear!".to_string());
     assert_eq!(received.author(), panda.id());
+}
+
+#[tokio::test]
+async fn replay_stream() {
+    setup_logging();
+
+    let chat_id = Topic::new();
+
+    let panda = p2panda::builder().spawn().await.unwrap();
+    let icebear = p2panda::builder().spawn().await.unwrap();
+
+    // Panda subscribes to chat and publishes one message.
+    {
+        let (mut panda_tx, _panda_rx) = panda.stream::<String>(chat_id).await.unwrap();
+        panda_tx.publish("Hello, Icebear!".into()).await.unwrap();
+    }
+
+    tokio::time::sleep(Duration::from_millis(100)).await;
+
+    // Panda subscribes again, this time asking to replay all messages.
+    let (_panda_tx, mut panda_rx) = panda
+        .stream_from::<String>(chat_id, Offset::Start)
+        .await
+        .unwrap();
+
+    // Icebear joins the chat and publishes one message.
+    let (mut icebear_tx, _icebear_rx) = icebear.stream::<String>(chat_id).await.unwrap();
+    icebear_tx.publish("Hello, Panda!".into()).await.unwrap();
+
+    // Panda should receive the first message they sent again, followed by Icebear's message which
+    // arrived via sync.
+    let mut received = vec![];
+
+    while let Some(event) = panda_rx.next().await {
+        if let StreamEvent::Processed(processed) = event {
+            received.push(processed);
+            if received.len() == 2 {
+                break;
+            }
+        }
+    }
+
+    assert_eq!(received[0].message(), &"Hello, Icebear!".to_string());
+    assert_eq!(received[1].message(), &"Hello, Panda!".to_string());
 }